// src/api/generated/models/ComputeJobStatus.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum ComputeJobStatus {
    NOT_STARTED = 'NotStarted',
    CREATED = 'Created',
    WAITING = 'Waiting',
    RUNNING = 'Running',
    SUCCEEDED = 'Succeeded',
    FAILED = 'Failed',
}